docs(README): automatic update
  Update README C#
  [This list might not be relevant]
  erb syntax

fixing erb syntax

Update radius documentation

docs(highlight): Add mention that only strings can be highlighted

Attributes cast as integer are still searched, but they are not
highlighted. This can result in confusing results in the dashboard and
integrations. I've made it explicit here.

I'm planning to write a longer guide on the subject as well. [ci skip]