feat: drop official support for .NET Framework 4.5
Following the discussions we had in the team, we have decided it was
time to officially drop the support for .NET Framework 4.5 which is an
old and outdated Windows-only binary target.

Since we are now targetting at least .NET Standard 1.3 or higher, the oldest
supported versions will now become.

 - .NET Framework 4.6.0 or higher (Windows only)
 - .NET Core 1.0.0 or higher (cross platform)

More detailled information regarding .NET implementations and support
can be found on the Microsoft documentation:

https://docs.microsoft.com/en-us/dotnet/standard/net-standard