fix(transport): MemoryStream disposed betweem two retryStrategy call (#748)
| Q                 | A
| ----------------- | ----------
| Bug fix?          | YES
| New feature?      | no
| BC breaks?        | no     
| Need Doc update   | no

## What?

Following on a expired certificate issue on one of `-dsn` domains, we noticed that in some cases, the C# client wasn't able to correctly retry a failing call. 

In that case the exception sent was:

```
System.Net.Http.HttpRequestException: The SSL connection could not be established, see inner exception. 
---> System.Security.Authentication.AuthenticationException: 
The remote certificate is invalid according to the validation procedure.
```

This SSL exception thrown by the `httpClient` was correctly catched and marked as a retryable failure by the strategy, but the following retry calls failed with the following exception:

```
System.Net.Http.HttpRequestException: Error while copying content to a stream.
 ---> System.ObjectDisposedException: Cannot access a closed Stream.
   at System.IO.StreamHelpers.ValidateCopyToArgs(Stream source, Stream destination, Int32 bufferSize)
``` 

This exception means that the `httpClient` wasn't able to read the stream that should be sent to the API. After some investigations I found out the reasons:

In the transport layer we create a `MemoryStream` **before** the retryStrategy loop:

https://github.com/algolia/algoliasearch-client-csharp/blob/c1f684e57eaee4b7c59874e5e5d1a9134714cc50/src/Algolia.Search/Transport/HttpTransport.cs#L116-L121

Then the retry calls the `SendRequest` method which does the following:

Gives the stream to `httpRequestMessage`:

https://github.com/algolia/algoliasearch-client-csharp/blob/c1f684e57eaee4b7c59874e5e5d1a9134714cc50/src/Algolia.Search/Http/AlgoliaHttpRequester.cs#L76-L81

Wrap the `httpRequestMessage` in a `using block`:

https://github.com/algolia/algoliasearch-client-csharp/blob/c1f684e57eaee4b7c59874e5e5d1a9134714cc50/src/Algolia.Search/Http/AlgoliaHttpRequester.cs#L93

And this is where the issue happens:

The stream created in the TransportLayer before the loop is **not copied** and thus `HttpRequestMessage` is taking ownership of it with the [new StreamContent()](https://docs.microsoft.com/en-us/dotnet/api/system.net.http.streamcontent.-ctor?view=net-5.0#System_Net_Http_StreamContent__ctor_System_IO_Stream_) ctor.

**Thus, at the end of the `using` block the stream will be closed/disposed and not available for further usages!** Which explains why the following calls failed. 

### Impact?

User using the client with only one `POST` method (methods having a stream/body to send like the search one) would have the call failed in loop because:

-> call n1 to `-dsn` with SSL expired. Retryable but dispose the `Stream`
-> call n2 to `-1` with the stream closed => throw a `HttpException` because of the closed stream => retryable
-> call n3 to `-2` with the stream closed => throw a `HttpException` because of the closed stream => retryable
-> call n4 to `-3` with the stream closed => throw a `HttpException` because of the closed stream => retryable

-> End with all the host marked as down..

On the other hand, if the `client` is used to target other endpoint such as `get` one, the get method will mark the `-dsn` as down for the other calls, making it possible for them to succeed. 

## How?

- Simply re-creating the Stream for each retryStrategy call instead of re-using it.

## Test strategy

- Added a `e2e` test for the retryStrategy checking that the client correctly retries and that the MemoryStream is available.
- Non regressions on existing tests

## Documentation

- https://docs.microsoft.com/en-us/dotnet/api/system.net.http.streamcontent.-ctor?view=net-5.0#System_Net_Http_StreamContent__ctor_System_IO_Stream_